pyproject.toml
[tool.poetry]
name = "TaskZenAPI"
version = "0.1.0"
description = "TaskZenAPI App"
authors = ["Islam Mukanov <mukanov_ir@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
drf-yasg = "^1.21.7"
djangorestframework = "^3.15.1"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
pytest = "^8.1.1"
pytest-django = "^4.8.0"
gunicorn = "^21.2.0"
django-debug-toolbar = "^4.3.0"
arrow = "^1.3.0"
django-filter = "^24.2"
django-cors-headers = "^4.3.1"
psycopg2 = "^2.9.9"
celery = {extras = ["redis"], version = "^5.4.0"}
flower = "^2.0.1"

[tool.poetry.dev-dependencies]
isort = "^5.13.2"
flake8 = "^7.0.0"
black = "^24.3.0"
mypy = "^1.9.0"

[tool.isort]
profile = "black"
multi_line_output = 3
force_grid_wrap = 0
combine_as_imports = true
skip_gitignore = true
skip_glob = ["**/migrations/*", "**/settings/*"]

[tool.black]
line-length = 90
target-version = ['py311']
extend-exclude = '''
(
  migrations   # Django DB migrations
)
'''


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"